task discrip

+++ argc_argv/4-add.c
@@ -1,6 +1,7 @@
 #include <stdio.h>
 #include <stdlib.h>
 #include <string.h>
+#include <ctype.h>
 /**
  * main - adds positive numbers.
  * @argc: number of command line arguments.
@@ -20,7 +21,7 @@ int get_number(int argc, char *argv[])
 	{
 		int j;
 
-		for (j = 0; j < strlen(argv[i]); j++)
+		for (j = 0; j < (int)strlen(argv[i]); j++)
 		{
 			if (!isdigit(argv[i][j]))
 			{
